Replace C-style casts with C++-style; remove vestiges of unmaintained Mac OS X code; add png format map; better cant-use messages for weapons and armor; use old-style dative case more consistently

+++ src/image.h
@@ -48,8 +48,8 @@ public:
     int x, y, width, height;
 };
 
-#define IM_OPAQUE (unsigned int)255
-#define IM_TRANSPARENT 0
+#define IM_OPAQUE 255u
+#define IM_TRANSPARENT 0u
 
 
 /**
